requirements.txt
streamlit>=1.28.0
faiss-cpu>=1.7.4
numpy>=1.24.0
pandas>=2.0.0
openai>=1.0.0
langchain-core>=0.1.0 